Allow more control over validating webhooks
The webhook builder doesn't currently allow for
much flexibility as it only allows you to define
the object being modified as the validator.

This change allows you to declare a separate
validator that can then do more without polluting
the API type.

In particular this also extends the interface
to provide the request, which grants access to
other information such as the user info.

It adds a wrapper that wraps a Validator in
the new interface to keep the number of code
paths small while maintaining compatibility
and the ease of just adding validation methods
to the api type for simple use cases.